Backend/requirements.txt
fastapi[all]
openai>=0.27.0
boto3>=1.26.0
SQLAlchemy>=2.0.0
PyMySQL>=1.0.2
pydantic-settings==2.1.0
pydantic[email]
email-validator>=2.0.0
httpx>=0.25.0
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-api-python-client
google-auth
google-cloud-aiplatform>=1.38.0
google-cloud-speech>=2.20.0
google-cloud-storage>=2.10.0
moviepy>=1.0.3
Pillow>=10.0.0
